place the config methods in their right folder
