+++ packages/pymodaq_gui/src/pymodaq_gui/config.py
@@ -1,6 +1,6 @@
 from pathlib import Path
 
-from pymodaq_utils.config import (GlobalConfig, BaseConfig)
+from pymodaq_utils.config import (GlobalConfig, BaseConfig, get_set_config_dir)
 
 
 @GlobalConfig.register()
@@ -9,3 +9,14 @@ class Config(BaseConfig):
     config_template_path = Path(__file__).parent.joinpath('resources/config_template.toml')
     config_name = f"gui"
 
+
+def get_set_layout_path(user=False):
+    """ creates and return the config folder path for layout files
+    """
+    return get_set_config_dir('layout_configs', user=user)
+
+
+def get_set_roi_path():
+    """ creates and return the config folder path for managers files
+    """
+    return get_set_config_dir('roi_configs')
